+++ liberty/utils/buckets.py
@@ -39,6 +39,15 @@ class S3Utils():
             logger.error(ce)
 
 
+    def key_exists(self, key):
+        try:
+            self.b2_client.head_object(Bucket=_ASSETS_BUCKET_NAME, Key=key)
+        except ClientError as e:
+            if e.response['Error']['Code'] == "404":
+                return False
+        return True
+
+
     # List the keys of the objects in the assets bucket
     async def list_object_keys(self):
         try:
@@ -53,10 +62,13 @@ class S3Utils():
             logger.error(ce)
 
 
-    async def upload_file(self, file_path, b2path=None):
+    async def upload_file(self, file_path, b2path=None, force=False):
         remote_path = b2path
         if remote_path is None:
             remote_path = file_path
+        if not force and self.key_exists(remote_path):
+            logger.warning(f'File {remote_path} already exists in bucket... Not Uploading...')
+            return
         logger.info(f'Uploading {file_path} to {remote_path}...')
         try:
             response = self.b2.Bucket(_ASSETS_BUCKET_NAME).upload_file(file_path, remote_path)
